+++ internal/provider/provider_test.go
@@ -105,17 +105,3 @@ provider "hcp" {}`)
 	c = append(c, res...)
 	return strings.Join(c, "\n")
 }
-
-func TestShouldRetryErrorCode(t *testing.T) {
-	errorCodesToRetry := []int{502, 503, 504}
-
-	shouldFail := shouldRetryErrorCode(200, errorCodesToRetry[:])
-	if shouldFail != false {
-		t.Errorf("shouldRetryErrorCode(200, []int{502, 503, 504}[:]) = %v; want false", shouldFail)
-	}
-
-	shouldSucceed := shouldRetryErrorCode(503, errorCodesToRetry[:])
-	if shouldSucceed != true {
-		t.Errorf("shouldRetryErrorCode(503, []int{502, 503, 504}[:]) = %v; want true", shouldSucceed)
-	}
-}
